fix last commit

+++ rtp.h
@@ -540,7 +540,7 @@ public:
 	void unlock() {
 		pthread_mutex_unlock(&mlock);
 	}
-	void update(uint32_t saddr, uint32_t time, uint8_t mosf1, uint8_t mosf2, uint8_t mosAD, uint16_t jitter, uint16_t loss);
+	void update(uint32_t saddr, uint32_t time, uint8_t mosf1, uint8_t mosf2, uint8_t mosAD, uint16_t jitter, double loss);
 	void flush_and_clean(map<uint32_t, node_t> *map);
 
 private:
